refactor: CompanySearchSpec 도입
