+++ company-service/src/main/java/com/firstlogistics/companyservice/domain/specification/CompanySearchSpec.java
@@ -1,10 +1,10 @@
-package com.firstlogistics.companyservice.domain.repository.dto;
+package com.firstlogistics.companyservice.domain.specification;
 
 import com.firstlogistics.companyservice.domain.enums.CompanyStatus;
 
 import java.util.UUID;
 
-public record CompanyQueryCondition(
+public record CompanySearchSpec(
         String keyword,
         String type,
         UUID hubId,
